librocketnpu: fix bias correction for int8 weights (wzp=0) + cleanup debug
Use signed int8 cast for weight bytes in calc_bias_correction to handle
both uint8 (wzp=128) and int8 (wzp=0) TFLite weight formats. Previous
unsigned subtraction gave wrong results for weight bytes >= 128 in int8
models. Also fix OUT_CVT offset cast and remove debug prints.

Note: YOLO output still saturates due to fundamental single-OUT_CVT
limitation — per-channel BS MUL works but max_conv_scale uniform OUT_CVT
overflows for high-ozp models. Needs per-channel OUT_CVT (multi-task)
approach matching RKNN.

+++ librocketnpu/src/rnpu_regcmd.c
@@ -1097,7 +1097,7 @@ static unsigned fill_brdma_per_channel_regcmd(const struct rnpu_model *model,
    }
 
    /* OUT_CVT: uniform scale from max_conv_scale */
-   unsigned offset = ozp - 0x80;
+   int offset = (int)ozp - 0x80;
    float conv_scale = (task->input_scale * task->weights_scale) / task->output_scale;
    uint32_t scale_bits = fui(conv_scale);
    unsigned shift = 127 + 31 - 32 - (scale_bits >> 23) + 16;
@@ -1105,7 +1105,7 @@ static unsigned fill_brdma_per_channel_regcmd(const struct rnpu_model *model,
    unsigned scale = ((scale_bits >> 9) & 0x7fff) + 1;
    if (scale < 1 << 14) scale |= 1 << 14;
 
-   EMIT(REG_DPU_OUT_CVT_OFFSET, offset);
+   EMIT(REG_DPU_OUT_CVT_OFFSET, (uint32_t)offset);
    EMIT(REG_DPU_OUT_CVT_SCALE, DPU_OUT_CVT_SCALE_OUT_CVT_SCALE(scale));
    EMIT(REG_DPU_OUT_CVT_SHIFT, DPU_OUT_CVT_SHIFT_OUT_CVT_SHIFT(shift - 1));
